requirements.txt
pandas
python-dotenv
langchain-groq
pdfplumber
pytest
tabulate